rename `.data._normalize` -> `.ticktools`
